type定義専用ファイルを作成して全てのtype定義を移動
