+++ src/App.tsx
@@ -2,11 +2,7 @@ import { Grid, Typography } from "@mui/material";
 import { InputMemo } from "./components/InputMemo";
 import { MemoList } from "./components/MemoList";
 import { Dispatch, createContext, useState } from "react";
-
-export type memoContextType = {
-  memoArray: string[];
-  setMemoArray: Dispatch<React.SetStateAction<string[]>>;
-};
+import { memoContextType } from "./type/type";
 
 export const memoContext = createContext<memoContextType>(
   {} as {
