+++ src/type/type.tsx
@@ -1,0 +1,26 @@
+import { Dispatch, SetStateAction } from "react";
+
+export type memoContextType = {
+  memoArray: string[];
+  setMemoArray: Dispatch<SetStateAction<string[]>>;
+};
+
+export type memoInput = {
+  memo: string;
+};
+
+export type MemoItemProps = {
+  memoItem: string;
+  memoIndex: number;
+};
+
+export type MemoEditDialogProps = {
+  memoEditDialogOpen: boolean;
+  memoEdtOnClose: () => void;
+  memoIndex: number;
+  setMemoEditDialogOpen: Dispatch<React.SetStateAction<boolean>>;
+};
+
+export type memoEdit = {
+  memoEdt: string;
+};
